sdk: fix tests and add changelog entry for interactive UDC withdraw
